examples/pipeline2/requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
DateTime==4.3
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
py4j==0.10.7
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
urllib3==1.25.7
Werkzeug==0.16.0
xgboost===1.0.0-SNAPSHOT
zope.interface==4.7.1
